Create fill input
