--- apps/studio/src/routes/editor/EditPanel/StylesTab/single/ColorInput/PopoverColorPicker.tsx
@@ -1,183 +1,0 @@
-import { useEditorEngine } from '@/components/Context';
-import styled from '@emotion/styled';
-import { checkPattern, ColorPicker } from '@onlook/ui/color-picker';
-import { Icons } from '@onlook/ui/icons';
-import {
-    Popover,
-    PopoverContent,
-    PopoverScrollArea,
-    PopoverSeparator,
-    PopoverTrigger,
-} from '@onlook/ui/popover';
-import { Color, isColorEmpty, type Palette } from '@onlook/utility';
-import { useEffect, useState } from 'react';
-import { twMerge } from 'tailwind-merge';
-
-const ColorButtonBackground = styled.div`
-    ${checkPattern('white', '#aaa', '8px')}
-`;
-
-const ColorButton: React.FC<
-    {
-        value?: Color;
-    } & React.PropsWithoutRef<JSX.IntrinsicElements['div']>
-> = ({ className, value, ...props }) => {
-    return (
-        <div
-            {...props}
-            className={twMerge(
-                'rounded-sm w-5 h-5 border border-white/20 cursor-pointer shadow bg-background',
-                className,
-            )}
-        >
-            {isColorEmpty(value?.toHex() ?? 'transparent') ? (
-                <div className="w-full h-full rounded-sm overflow-hidden bg-background-secondary"></div>
-            ) : (
-                <ColorButtonBackground className="w-full h-full rounded-sm overflow-hidden">
-                    <div
-                        className="w-full h-full rounded-[1.5px]"
-                        style={{
-                            backgroundColor: value?.toHex() ?? 'transparent',
-                        }}
-                    />
-                </ColorButtonBackground>
-            )}
-        </div>
-    );
-};
-
-interface PopoverPickerProps {
-    color: Color;
-    onChange: (color: Color) => void;
-    onChangeEnd: (color: Color) => void;
-}
-
-export const PopoverPicker = ({ color, onChange, onChangeEnd }: PopoverPickerProps) => {
-    const editorEngine = useEditorEngine();
-
-    const [isOpen, toggleOpen] = useState(false);
-    const [viewMode, setViewMode] = useState<'grid' | 'list'>('grid');
-    const [palette, setPalette] = useState<Palette>(color.palette);
-
-    useEffect(() => {
-        if (isOpen && !editorEngine.history.isInTransaction) {
-            editorEngine.history.startTransaction();
-            return;
-        }
-
-        if (!isOpen && editorEngine.history.isInTransaction) {
-            editorEngine.history.commitTransaction();
-        }
-        return () => editorEngine.history.commitTransaction();
-    }, [isOpen]);
-
-    useEffect(() => {
-        setPalette(color.palette);
-    }, [color]);
-
-    function renderColorPicker() {
-        return (
-            <ColorPicker
-                color={color}
-                onMouseDown={() => editorEngine.history.startTransaction()}
-                onChange={onChange}
-                onChangeEnd={(val) => {
-                    onChangeEnd?.(val);
-                    setPalette(val.palette);
-                }}
-            />
-        );
-    }
-
-    function renderPalette() {
-        const colors = Object.keys(palette.colors);
-        return (
-            <div className="px-0.5 py-1.5">
-                {viewMode === 'grid' ? (
-                    <div className="grid grid-cols-7 gap-1.5 p-1 text-center justify-between">
-                        {colors.map((level) => (
-                            <div
-                                key={level}
-                                className="w-6 h-6 content-center cursor-pointer rounded border-[0.5px] border-foreground-tertiary/50"
-                                style={{ backgroundColor: palette.colors[Number.parseInt(level)] }}
-                                onClick={() =>
-                                    onChangeEnd(Color.from(palette.colors[Number.parseInt(level)]))
-                                }
-                            >
-                                {/* Commenting out so that we can use this for tooltips over these grid elements */}
-                                {/* <p
-                                    className={cn(
-                                        'text-xs text-white drop-shadow-lg',
-                                        parseInt(level) < 500 ? 'invert' : '',
-                                    )}
-                                > */}
-                            </div>
-                        ))}
-                    </div>
-                ) : (
-                    <div className="flex flex-col">
-                        {colors.map((level) => (
-                            <div
-                                className="gap-2 hover:bg-background-secondary p-1 flex align-center cursor-pointer rounded-md group"
-                                key={level}
-                                onClick={() =>
-                                    onChangeEnd(Color.from(palette.colors[Number.parseInt(level)]))
-                                }
-                            >
-                                <div
-                                    key={level}
-                                    className="w-5 h-5 content-center rounded border-[0.5px] border-foreground-tertiary/50"
-                                    style={{
-                                        backgroundColor: palette.colors[Number.parseInt(level)],
-                                    }}
-                                />
-                                <div className="text-small text-foreground-secondary group-hover:text-foreground-primary">
-                                    <span>
-                                        {palette.name}-{level}
-                                    </span>
-                                </div>
-                            </div>
-                        ))}
-                    </div>
-                )}
-            </div>
-        );
-    }
-
-    return (
-        <Popover onOpenChange={(open) => toggleOpen(open)}>
-            <PopoverTrigger>
-                <ColorButton value={color} onClick={() => toggleOpen(!isOpen)} />
-            </PopoverTrigger>
-            <PopoverContent
-                align="end"
-                className="min-w-12 bg-background/90 backdrop-blur-lg z-10 rounded-lg p-0 shadow-xl overflow-hidden"
-            >
-                <div className="flex flex-col justify-between items-center">
-                    {renderColorPicker()}
-                    <PopoverSeparator />
-                    <div className="flex flex-row items-center justify-between w-full px-2.5 py-1.5">
-                        <span className="text-foreground-secondary text-smallPlus">
-                            {palette.name}
-                        </span>
-                        <button
-                            aria-label={`Toggle ${viewMode === 'grid' ? 'list' : 'grid'} mode`}
-                            className="text-foreground-tertiary hover:text-foreground-hover rounded"
-                            onClick={() => {
-                                setViewMode(viewMode === 'grid' ? 'list' : 'grid');
-                            }}
-                        >
-                            {viewMode === 'grid' ? (
-                                <Icons.ViewGrid className="h-4 w-4" />
-                            ) : (
-                                <Icons.ViewHorizontal className="h-4 w-4" />
-                            )}
-                        </button>
-                    </div>
-                    <PopoverSeparator />
-                    <PopoverScrollArea className="h-32 px-1">{renderPalette()}</PopoverScrollArea>
-                </div>
-            </PopoverContent>
-        </Popover>
-    );
-};
